# Made sure visualisation functions include periodic copies of particles (attempt 2)