--- a/swiftsimio/visualisation/projection_backends/fast.py
+++ b/swiftsimio/visualisation/projection_backends/fast.py
@@ -24,7 +24,15 @@ from swiftsimio.visualisation.projection_backends.kernels import (
 
 
 @jit(nopython=True, fastmath=True)
-def scatter(x: float64, y: float64, m: float32, h: float32, res: int) -> ndarray:
+def scatter(
+    x: float64,
+    y: float64,
+    m: float32,
+    h: float32,
+    res: int,
+    box_x: float64,
+    box_y: float64,
+) -> ndarray:
     """
     Creates a weighted scatter plot
 
@@ -84,68 +92,90 @@ def scatter(x: float64, y: float64, m: float32, h: float32, res: int) -> ndarray
     # Pre-calculate this constant for use with the above
     inverse_cell_area = res * res
 
-    for x_pos, y_pos, mass, hsml in zip(x, y, m, h):
-        # Calculate the cell that this particle; use the 64 bit version of the
-        # resolution as this is the same type as the positions
-        particle_cell_x = int32(float_res_64 * x_pos)
-        particle_cell_y = int32(float_res_64 * y_pos)
+    for x_pos_original, y_pos_original, mass, hsml in zip(x, y, m, h):
+        for xshift in range(3):
+            for yshift in range(3):
+                x_pos = x_pos_original + (xshift - 1) * box_x
+                y_pos = y_pos_original + (yshift - 1) * box_y
 
-        # SWIFT stores hsml as the FWHM.
-        kernel_width = kernel_gamma * hsml
+                # Calculate the cell that this particle; use the 64 bit version of the
+                # resolution as this is the same type as the positions
+                particle_cell_x = int32(float_res_64 * x_pos)
+                particle_cell_y = int32(float_res_64 * y_pos)
 
-        # The number of cells that this kernel spans
-        cells_spanned = int32(1.0 + kernel_width * float_res)
+                # SWIFT stores hsml as the FWHM.
+                kernel_width = kernel_gamma * hsml
 
-        if (
-            particle_cell_x + cells_spanned < 0
-            or particle_cell_x - cells_spanned > maximal_array_index
-            or particle_cell_y + cells_spanned < 0
-            or particle_cell_y - cells_spanned > maximal_array_index
-        ):
-            # Can happily skip this particle
-            continue
+                # The number of cells that this kernel spans
+                cells_spanned = int32(1.0 + kernel_width * float_res)
 
-        if cells_spanned <= 1:
-            # Easygame, gg
-            if (
-                particle_cell_x >= 0
-                and particle_cell_x <= maximal_array_index
-                and particle_cell_y >= 0
-                and particle_cell_y <= maximal_array_index
-            ):
-                image[particle_cell_x, particle_cell_y] += mass * inverse_cell_area
-        else:
-            # Now we loop over the square of cells that the kernel lives in
-            for cell_x in range(
-                # Ensure that the lowest x value is 0, otherwise we'll segfault
-                max(0, particle_cell_x - cells_spanned),
-                # Ensure that the highest x value lies within the array bounds,
-                # otherwise we'll segfault (oops).
-                min(particle_cell_x + cells_spanned + 1, maximal_array_index + 1),
-            ):
-                # The distance in x to our new favourite cell -- remember that our x, y
-                # are all in a box of [0, 1]; calculate the distance to the cell centre
-                distance_x = (float32(cell_x) + 0.5) * pixel_width - float32(x_pos)
-                distance_x_2 = distance_x * distance_x
-                for cell_y in range(
-                    max(0, particle_cell_y - cells_spanned),
-                    min(particle_cell_y + cells_spanned + 1, maximal_array_index + 1),
+                if (
+                    particle_cell_x + cells_spanned < 0
+                    or particle_cell_x - cells_spanned > maximal_array_index
+                    or particle_cell_y + cells_spanned < 0
+                    or particle_cell_y - cells_spanned > maximal_array_index
                 ):
-                    distance_y = (float32(cell_y) + 0.5) * pixel_width - float32(y_pos)
-                    distance_y_2 = distance_y * distance_y
+                    # Can happily skip this particle
+                    continue
 
-                    r = sqrt(distance_x_2 + distance_y_2)
+                if cells_spanned <= 1:
+                    # Easygame, gg
+                    if (
+                        particle_cell_x >= 0
+                        and particle_cell_x <= maximal_array_index
+                        and particle_cell_y >= 0
+                        and particle_cell_y <= maximal_array_index
+                    ):
+                        image[particle_cell_x, particle_cell_y] += (
+                            mass * inverse_cell_area
+                        )
+                else:
+                    # Now we loop over the square of cells that the kernel lives in
+                    for cell_x in range(
+                        # Ensure that the lowest x value is 0, otherwise we'll segfault
+                        max(0, particle_cell_x - cells_spanned),
+                        # Ensure that the highest x value lies within the array bounds,
+                        # otherwise we'll segfault (oops).
+                        min(
+                            particle_cell_x + cells_spanned + 1, maximal_array_index + 1
+                        ),
+                    ):
+                        # The distance in x to our new favourite cell -- remember that our x, y
+                        # are all in a box of [0, 1]; calculate the distance to the cell centre
+                        distance_x = (float32(cell_x) + 0.5) * pixel_width - float32(
+                            x_pos
+                        )
+                        distance_x_2 = distance_x * distance_x
+                        for cell_y in range(
+                            max(0, particle_cell_y - cells_spanned),
+                            min(
+                                particle_cell_y + cells_spanned + 1,
+                                maximal_array_index + 1,
+                            ),
+                        ):
+                            distance_y = (
+                                float32(cell_y) + 0.5
+                            ) * pixel_width - float32(y_pos)
+                            distance_y_2 = distance_y * distance_y
 
-                    kernel_eval = kernel(r, kernel_width)
+                            r = sqrt(distance_x_2 + distance_y_2)
 
-                    image[cell_x, cell_y] += mass * kernel_eval
+                            kernel_eval = kernel(r, kernel_width)
+
+                            image[cell_x, cell_y] += mass * kernel_eval
 
     return image
 
 
 @jit(nopython=True, fastmath=True, parallel=True)
 def scatter_parallel(
-    x: float64, y: float64, m: float32, h: float32, res: int
+    x: float64,
+    y: float64,
+    m: float32,
+    h: float32,
+    res: int,
+    box_x: float64,
+    box_y: float64,
 ) -> ndarray:
     """
     Parallel implementation of scatter
@@ -216,6 +246,8 @@ def scatter_parallel(
             m=m[left_edge:right_edge],
             h=h[left_edge:right_edge],
             res=res,
+            box_x=box_x,
+            box_y=box_y,
         )
 
     return output