--- a/swiftsimio/visualisation/projection_backends/histogram.py
+++ b/swiftsimio/visualisation/projection_backends/histogram.py
@@ -27,7 +27,7 @@ def scatter(
 
     Computes contributions to from particles with positions
     (`x`,`y`) with smoothing lengths `h` weighted by quantities `m`.
-    This ignores boundary effects.
+    This includes periodic boundary effects.
 
     Parameters
     ----------
@@ -47,6 +47,14 @@ def scatter(
     res : int
         the number of pixels along one axis, i.e. this returns a square
         of res * res.
+
+    box_x: float64
+        box size in x, in the same rescaled length units as x and y. Used
+        for periodic wrapping.
+
+    box_y: float64
+        box size in y, in the same rescaled length units as x and y. Used
+        for periodic wrapping.
 
     Returns
     -------
@@ -78,6 +86,7 @@ def scatter(
     inverse_cell_area = float_res * float_res
 
     for x_pos_original, y_pos_original, mass in zip(x, y, m):
+        # loop over periodic copies of this particle
         for xshift in range(3):
             for yshift in range(3):
                 x_pos = x_pos_original + (xshift - 1) * box_y
@@ -117,7 +126,7 @@ def scatter_parallel(
     Creates a weighted scatter plot. Computes contributions from
     particles with positions (`x`,`y`) with smoothing lengths `h`
     weighted by quantities `m`.
-    This ignores boundary effects.
+    This includes periodic boundary effects.
 
     Parameters
     ----------
@@ -136,6 +145,14 @@ def scatter_parallel(
     res : int
         the number of pixels along one axis, i.e. this returns a square
         of res * res.
+
+    box_x: float64
+        box size in x, in the same rescaled length units as x and y. Used
+        for periodic wrapping.
+
+    box_y: float64
+        box size in y, in the same rescaled length units as x and y. Used
+        for periodic wrapping.
 
     Returns
     -------