… docs/source/visualisation/projection.rst
… docs/source/visualisation/slice.rst
… docs/source/visualisation/volume_render.rst
… swiftsimio/visualisation/projection_backends/renormalised.py
… swiftsimio/visualisation/projection_backends/subsampled.py
… swiftsimio/visualisation/projection_backends/subsampled_extreme.py
… swiftsimio/visualisation/volume_render.py
… tests/test_visualisation.py