New version of #148 that implements the periodic wrapping on the backend level. The memory footprint is not affected by this version, but there should still be a (small) performance penalty.

Here are some timings for my small test based on `cosmological_volume.hdf5`:

Old version:
```
Projection:
Backend "histogram" took 743.87 +- 12.995156135586896 ms
Backend "fast" took 620.74 +- 32.442475264777684 ms
Backend "renormalised" took 1128.18 +- 201.63854249894925 ms
Backend "subsampled" took 2442.19 +- 135.6247984160516 ms
Backend "subsampled_extreme" took 2361.57 +- 158.01341649057622 ms
Backend "reference" took 1250.00 +- 166.40818857369823 ms
Backend "gpu" took 41.47 +- 10.81907894761692 ms
Slicing:
Took 419.78 +- 8.391963275394627 ms
Volume rendering:
Took 4440.49 +- 59.48874302494967 ms
```

New version:
```
Projection:
Backend "histogram" took 733.99 +- 33.32694376915969 ms
Backend "fast" took 584.81 +- 20.463084650768796 ms
Backend "renormalised" took 785.32 +- 20.98768985286519 ms
Backend "subsampled" took 1270.11 +- 28.538669311331958 ms
Backend "subsampled_extreme" took 1263.11 +- 44.949929481954534 ms
Backend "reference" took 756.16 +- 27.871183567513178 ms
Backend "gpu" took 53.32 +- 13.414277270534988 ms
Slicing:
Took 411.87 +- 8.084733606619707 ms
Volume rendering:
Took 4338.15 +- 241.85104295403843 ms
```

For some reason all of them (except for the gpu backend) were actually faster in the new version. Not sure I understand why.

I am still running a larger and more representative test.