@bwvdnbro 2022-09-15
I have tested this on a 20 Mpc thick slab of a 1Gpc FLAMINGO simulation. All backends except for `reference` (which spits out lots of "Reference grid not created at a high enough resolution" messages and actually gives the wrong result) are faster in the new version, despite doing a bit more work. The results are the same as with the old version.