@JBorrow 2022-09-19
Reference should do that, it's designed to make sure that every particle's kernel is well resolved (if you reach a particle that is not resolved, it quits - hence the wrong answer).

The other stuff looks great, thank you. I am concerned that we are changing an API that is used by external code here, though. Is there some way that we can make box_x and box_y optional? Is there a performance impact if they are set to None in the case where you don't want periodicity?