@bwvdnbro 2022-09-21
Numba does not look `Union[float64, None]`, so I have made the default `box_x/y=0.`, which actually makes sense as well. I have now also added an optional argument `periodic = True` to all visualisation functions. The non-periodic version is slightly faster, as expected.

Do you think it would be worth adding something about periodic boundaries to the documentation? I had a quick look, but I feel that this does not really fit well within the current text.