@JBorrow 2022-09-21
I don't think we need to describe the general concept of periodic boundaries (we should leave that to the SWIFT docs) but we'll need to update the docs with the new function arguments and describing the general concept of projecting over periodic boundaries.