@JBorrow 2022-09-23
Final question: does this use the periodic flag to see if the projection should be done periodically?