@bwvdnbro 2022-09-26
No, it does not. It also took me a while to find a "periodic flag". The only place where the periodic flag is stored is in the parameter list in the snapshot, as `InitialConditions:periodic`. I could try to read that, but that is not very clean. Is there another "periodic flag" I am not aware of?