@bwvdnbro 2022-10-27
@JBorrow any thoughts about my last comment?

The default behaviour now is to assume periodic boundaries if not specified, which seems sensible for most actual applications. The default for the backends themselves is to assume no periodic boundaries, since those do not automatically have access to the box size.