@JBorrow 2022-10-27
Okay, that's fine then. Let's keep it that way, but make sure to document this behavior choice thoroughly.

Have you checked for correctness? I suppose there are the automated tests, but are there periodic correctness tests?