@bwvdnbro 2022-10-27
I have done the tests described in the previous version of this PR (#148). Not sure if there is something we can do automatically to test this.