@JBorrow 2022-10-27
Is it possible to construct one large kernel that overlaps the boundary and check that the answer is the same in periodic mode and adding two non-periodic images together?

E.g. something like this:

```
-          -
---      ---
----    ----
---      ---
-          -
```